+++ BambooTracker/gui/mainwindow.cpp
@@ -3123,12 +3123,8 @@ void MainWindow::on_actionNew_triggered()
 		case QMessageBox::Yes:
 			if (!on_actionSave_triggered()) return;
 			break;
-		case QMessageBox::No:
-			break;
-		case QMessageBox::Cancel:
-			return;
 		default:
-			break;
+			return;
 		}
 	}